+++ install/installCyberPanel.py
@@ -785,16 +785,16 @@ module cyberpanel_ols {
                 self.install_package(f'lsphp{version}*', '--skip-broken')
                 
         elif self.distro == cent8:
-            # Install PHP versions in batches with exclusions
-            exclude_flags = "--exclude lsphp73-pecl-zip --exclude *imagick*"
-            
-            # First batch: PHP 7.x and 8.0
-            versions_batch1 = ' '.join([f'lsphp{v}*' for v in php_versions[:5]])
-            self.install_package(versions_batch1, f'{exclude_flags} --skip-broken')
-            
-            # Second batch: PHP 8.1+
-            versions_batch2 = ' '.join([f'lsphp{v}*' for v in php_versions[5:]])
-            self.install_package(versions_batch2, f'{exclude_flags} --skip-broken')
+            install_utils.ensure_lsphp_runtime_deps(self.distro)
+            exclude_flags = (
+                "--exclude lsphp73-pecl-zip --exclude *imagick* "
+                "--exclude 'lsphp*-imap' --exclude 'lsphp*-mbstring'"
+            )
+            all_versions = ' '.join([f'lsphp{v}*' for v in php_versions])
+            self.install_package(
+                all_versions,
+                f'{exclude_flags} --skip-broken --nobest',
+            )
             
         elif self.distro == openeuler:
             # Install all PHP versions at once
@@ -1092,30 +1092,66 @@ gpgcheck=1
 
         install_utils.call(command, self.distro, command, command, 1, 1, os.EX_OSERR, True)
 
+        install_utils.ensure_mariadb_client_cli(self.distro)
+
         ############## Start mariadb ######################
 
         self.startMariaDB()
 
     def changeMYSQLRootPassword(self):
-        if self.remotemysql == 'OFF':
-            if self.distro == ubuntu:
-                passwordCMD = "use mysql;DROP DATABASE IF EXISTS test;DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%%';GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost' IDENTIFIED BY '%s';UPDATE user SET plugin='' WHERE User='root';flush privileges;" % (
-                    InstallCyberPanel.mysql_Root_password)
-            else:
-                passwordCMD = "use mysql;DROP DATABASE IF EXISTS test;DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%%';GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost' IDENTIFIED BY '%s';flush privileges;" % (
-                    InstallCyberPanel.mysql_Root_password)
+        if self.remotemysql != 'OFF':
+            return
 
-            # For AlmaLinux 9, try mysql command first, then mariadb
-            if self.distro == cent8 or self.distro == openeuler:
-                command = 'mysql -u root -e "' + passwordCMD + '"'
-                result = install_utils.call(command, self.distro, command, command, 0, 0, os.EX_OSERR)
-                if result != 0:
-                    # If mysql command fails, try mariadb
-                    command = 'mariadb -u root -e "' + passwordCMD + '"'
-                    install_utils.call(command, self.distro, command, command, 0, 0, os.EX_OSERR)
-            else:
-                command = 'mariadb -u root -e "' + passwordCMD + '"'
-                install_utils.call(command, self.distro, command, command, 0, 0, os.EX_OSERR)
+        install_utils.ensure_mariadb_client_cli(self.distro)
+        cli = install_utils.resolve_mysql_cli()
+        if not cli:
+            self.stdOut(
+                "Error: mysql/mariadb client not found after MariaDB install.",
+                0,
+            )
+            return
+
+        passwordCMD = (
+            "use mysql;DROP DATABASE IF EXISTS test;"
+            "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%%';"
+            "ALTER USER 'root'@'localhost' IDENTIFIED BY '%s';"
+            "GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost' WITH GRANT OPTION;"
+            "flush privileges;"
+        ) % (InstallCyberPanel.mysql_Root_password)
+
+        if self.distro == ubuntu:
+            passwordCMD = (
+                "use mysql;DROP DATABASE IF EXISTS test;"
+                "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%%';"
+                "ALTER USER 'root'@'localhost' IDENTIFIED BY '%s';"
+                "UPDATE user SET plugin='' WHERE User='root';"
+                "flush privileges;"
+            ) % (InstallCyberPanel.mysql_Root_password)
+
+        socket_cmds = [
+            'sudo %s' % cli,
+            'sudo /usr/bin/mariadb',
+            'sudo /usr/bin/mysql',
+        ]
+        password_cmds = [
+            '%s -u root' % cli,
+            'mariadb -u root',
+            'mysql -u root',
+        ]
+
+        for cmd in socket_cmds + password_cmds:
+            command = '%s -e "%s"' % (cmd, passwordCMD)
+            result = install_utils.call(
+                command, self.distro, command, command, 0, 0, os.EX_OSERR,
+            )
+            if result == 0:
+                self.stdOut("MySQL root password set using: %s" % cmd, 1)
+                return
+
+        self.stdOut(
+            "Failed to set MySQL root password; configure manually if needed.",
+            0,
+        )
 
     def startMariaDB(self):